requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pandas>=2.1.3
numpy>=1.26.2
scikit-learn>=1.3.2
elasticsearch>=8.11.0
python-dotenv>=1.0.0
requests>=2.31.0
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-exporter-otlp>=1.21.0
python-jose>=3.3.0
aiohttp>=3.9.1
boto3>=1.33.6
python-multipart>=0.0.6
